feat(client): add `Toc` side panel (#612)
style: rename `useMenuFile` to `storeElectronMenu`
perf: change instance 'Cherry' mode
