+++ client/src/store/storeElectronMenu.ts
@@ -5,15 +5,14 @@ import { defineStore } from 'pinia'
  * @param isTextChange - 文本是否被改变
  * @param saveFilePath - 保存文件的路径
  */
-type MenuFileState = {
+type ElectronMenuState = {
   isTextChange: boolean;
   saveFilePath: string
 }
 
-export const useMenuFileStore = defineStore('elMenu', {
-  state: (): MenuFileState => ({
+export const useStoreElectronMenu = defineStore('el-menu', {
+  state: (): ElectronMenuState => ({
     isTextChange: false,
     saveFilePath: '',
   }),
-  
-})
+})
